refactor: extract guest checkout to its own component and edit checkout accordingly
